+++ src/lib/forms-schema/index.ts
@@ -712,6 +712,7 @@ const approvalStepNodeProps = {
 }
 
 const formSchema = Joi.object().keys({
+  workspaceId: Joi.number(),
   formsAppEnvironmentId: Joi.number().required(),
   name: Joi.string().required(),
   description: Joi.string().allow('', null),